Add sync_thread_has_updates util
